pyproject.toml
[tool.poetry]
name = "gltf-hub"
version = "0.0.1"
description = ""
authors = ["Benjamin Wasty <benny.wasty@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.0.3"
psycopg2 = "^2.8.4"
djangorestframework = "^3.11.0"
uvicorn = "^0.11.3"
whitenoise = {version = "^5.0.1", extras = ["brotli"] }
drf-yasg = "^1.17.1"
Pillow = "^7.0.0"

[tool.poetry.dev-dependencies]
coverage = "^5.0.3"
PyYAML = "^5.3"
profilehooks = "^1.11.1"
django-debug-toolbar = "^2.2"
invoke = "^1.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
